requirements.txt
pyserial>=3.5
flask>=3.1
